feat : "History Controller Test 추가"
